win-prompt/requirements.txt
google-generativeai
fastapi
uvicorn
python-dotenv
requests
pydantic
python-multipart
